added project D2RImageManipulation to enforce better separation of converns
